backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.8.0
httpx==0.27.0
python-dotenv==1.0.1
